Se actualiza REPO
Se traslada Backend al repo correspondiente

+++ ElBuenSabor/src/main/java/com/powerRanger/ElBuenSabor/dtos/FacturaCreateRequestDTO.java
@@ -8,17 +8,17 @@ public class FacturaCreateRequestDTO {
 
     // Otros campos opcionales que podrían venir al momento de generar la factura,
     // como datos de MercadoPago, si no se obtienen directamente del pedido o de otra fuente.
-    private Integer mpPaymentId;
-    private Integer mpMerchantOrderId;
+    private Long mpPaymentId;
+    private Long mpMerchantOrderId;
     private String mpPreferenceId;
     private String mpPaymentType;
 
     public Integer getPedidoId() { return pedidoId; }
     public void setPedidoId(Integer pedidoId) { this.pedidoId = pedidoId; }
-    public Integer getMpPaymentId() { return mpPaymentId; }
-    public void setMpPaymentId(Integer mpPaymentId) { this.mpPaymentId = mpPaymentId; }
-    public Integer getMpMerchantOrderId() { return mpMerchantOrderId; }
-    public void setMpMerchantOrderId(Integer mpMerchantOrderId) { this.mpMerchantOrderId = mpMerchantOrderId; }
+    public Long getMpPaymentId() { return mpPaymentId; }
+    public void setMpPaymentId(Long mpPaymentId) { this.mpPaymentId = mpPaymentId; }
+    public Long getMpMerchantOrderId() { return mpMerchantOrderId; }
+    public void setMpMerchantOrderId(Long mpMerchantOrderId) { this.mpMerchantOrderId = mpMerchantOrderId; }
     public String getMpPreferenceId() { return mpPreferenceId; }
     public void setMpPreferenceId(String mpPreferenceId) { this.mpPreferenceId = mpPreferenceId; }
     public String getMpPaymentType() { return mpPaymentType; }
